requirements.txt
django
djangorestframework
python-decouple
django-cors-headers
requests